Add full read and test support
Compleetly rewrote the read method to be fully async.
Update the comments on the database ts file.
Add the ability to run tests on the project.
Retarget the project to es6 for async support.
Change folder structure for a more intuitive layout.
Add a seperate launch type for VS Code for the ability to run tests on the project.