Add type checking and comments
Ad type checking to the query so that when documents are searched by _ID that it would automatically change the type to a ObjectID instead of a string.
Add more comments to the code so that it is easier to understand.
Remove unnecessary code from the close method.
Made the read test in the test script more dynamic by setting the query to read from the main object for its query settings.
Add the domain name setting to the API script in anticipation for adding domain name support to the API after the database interface has been completed.