Finish the Database Interface
Add spell checker for "teh" and "hte". A Must have for me.
Remove error checking code for DB connection status in each of the database operations as it already exists in teh connection section itself and would just be repeating itself.
Add write, update and delete operations to the interface.
Teasted each, code is working.
Add types to existing database operations.
Add comments everywhere.

Added intial test to the test script. This is currently fundimentially flawed and testing should not be used at this time as they all execute at the same time and some of the values are hard coded for certain tests. This was to test the functionality of the database interface itself while it was being created.
The next update should fix the test script.