Switch to Express for API server
Switch project to Express instead of Restify.
Http verbs now work properly, e.g. you can send a http delete request or a patch request instead of having to send post requests to diffrent URLs.
This move was also to prep for PassportJS.
Load PassportJS as an addon for express with the local module