Add more options
Add URL for users to be redirected to if they are not logged in.
Add URL to redirect users to if they log in successfully.
Add URL to redirect users to if they are unable to log in successfully.
Remove name option form the session config as it is automatically defined normally and does not need to be overriden.