Prep for user sorting
Add the prerequisites for user sorting to the database schema by adding event children as a seperate child type. (will revert this as it is the wrong move)
Fix a bug in the caegory schema where number was "Number" instead of "number", those are two diffrent data types.
Add missing metadata to getAllCategories.
Add the ability to get one category by specifying its ID.
Mark remove category for fixing later as it needs updating.