Rebuild user sorting
Rebuilt the user sorting to be easier to understand and more flexible:
Added the ability to label the category type.
Added the ability to store the enabled or disabled state of a category.
Remove eventChildren schema option.
Compleetly rewrite the user sorter function.
Update the category deleter function to use the new schema.
TODO:
Build the event user sorter and update the typings of the user sorter.